#164 Add creation of the welcome pages based on metadata structure. First test of composition of the document in svelte (with success)

+++ packages/backend/test/services/mongodb-media.test.ts
@@ -67,9 +67,10 @@ describe('Media service', () => {
   })
 
   beforeEach(async () => {
-    fs.rmdirSync('test/uploads', {
-      recursive:true
-    });
+    if(fs.existsSync('test/uploads'))
+      fs.rmdirSync('test/uploads', {
+        recursive:true
+      });
     await initMongoDbTestDatabase();
     params.route = {};
   })
